30.3

+++ kodilla-patterns2/src/main/java/com/kodilla/patterns2/adapter/bookclasifier/MedianAdaptee.java
@@ -9,13 +9,13 @@ import java.util.Map;
 
 public class MedianAdaptee implements BookStatistics {
     @Override
-    public int averagePublicationYear(Map<BookSignature, Book> books) {
+    public double averagePublicationYear(Map<BookSignature, Book> books) {
         Statistics statistics = new Statistics();
         return statistics.averagePublicationYear(books);
     }
 
     @Override
-    public int medianPublicationYear(Map<BookSignature, Book> books) {
+    public double medianPublicationYear(Map<BookSignature, Book> books) {
         Statistics statistics = new Statistics();
         return statistics.medianPublicationYear(books);
     }
