+++ kodilla-patterns2/src/main/java/com/kodilla/patterns2/adapter/bookclasifier/MedianAdapter.java
@@ -10,7 +10,7 @@ import java.util.Set;
 
 public class MedianAdapter extends MedianAdaptee  implements Classifier {
     @Override
-    public int publicationYearMedian(Set<Book> bookSet) {
+    public double publicationYearMedian(Set<Book> bookSet) {
 
         Map<BookSignature, com.kodilla.patterns2.adapter.bookclasifier.libraryb.Book> books = new HashMap<>();
